refactor(cli,linter): move LintOptions from cli to Linter
