+++ crates/oxc_cli/src/runner.rs
@@ -3,26 +3,30 @@ use std::{
     process::{ExitCode, Termination},
 };
 
-use clap::Command;
+use clap::{ArgMatches, Command};
 
-pub trait RunnerOptions: for<'a> From<&'a clap::ArgMatches> {
-    /// Add CLI arguments to a [`Command`], usually created from a [`Runner`].
-    fn build_args(cmd: Command) -> Command;
-}
-
-/// A trait for exposing Oxide functionality to the CLI.
+/// A trait for exposing functionality to the CLI.
 pub trait Runner: Send + Sync {
     /// The name of the runner. Used to create a subcommand.
     const NAME: &'static str;
+
     /// A short description about what the runner does
     const ABOUT: &'static str;
-    type Options: RunnerOptions;
 
     fn command() -> Command {
-        let cmd = Command::new(Self::NAME).about(Self::ABOUT);
-        Self::Options::build_args(cmd)
+        let mut command = Self::init_command();
+        if command.get_name().is_empty() {
+            command = command.name(Self::NAME);
+        }
+        if command.get_about().is_none() {
+            command = command.about(Self::ABOUT);
+        }
+        command
     }
-    fn new(options: Self::Options) -> Self;
+
+    fn new(matches: &ArgMatches) -> Self;
+
+    fn init_command() -> Command;
 
     /// Executes the runner, providing some result to the CLI.
     fn run(&self) -> CliRunResult;
@@ -137,20 +141,17 @@ mod tests {
         }
     }
 
-    impl RunnerOptions for TestRunnerOptions {
-        fn build_args(cmd: Command) -> Command {
-            cmd.arg(Arg::new("foo").short('f').action(ArgAction::SetTrue).required(false))
-        }
-    }
-
     impl Runner for TestRunner {
-        type Options = TestRunnerOptions;
-
         const ABOUT: &'static str = "some description";
         const NAME: &'static str = "test";
 
-        fn new(options: Self::Options) -> Self {
-            Self { options }
+        fn init_command() -> Command {
+            Command::new("")
+                .arg(Arg::new("foo").short('f').action(ArgAction::SetTrue).required(false))
+        }
+
+        fn new(matches: &ArgMatches) -> Self {
+            Self { options: TestRunnerOptions::from(matches) }
         }
 
         fn run(&self) -> CliRunResult {
@@ -167,8 +168,7 @@ mod tests {
     fn smoke_test_runner() {
         let cmd = TestRunner::command();
         let matches = cmd.get_matches_from("test -f".split(' '));
-        let opts = TestRunnerOptions::from(&matches);
-        let runner = TestRunner::new(opts);
+        let runner = TestRunner::new(&matches);
         assert!(runner.options.foo);
         let result = runner.run();
         assert!(matches!(result, CliRunResult::None));
